changed default implementation to standard view selection
The default installation just reads from the contained view and builds
the search data base from the results. This will take some seconds at
the first startup.
To speed things up, an alternative version uses a Shared Object that
holds the raw data in a compressed table. As this need a defined Shared
Object and there seems to be no SO plugin for SAPLink yet, the default
is changed to the direct read from the data base.