Got rid of all the warnings
I removed all the warnings in the Eclipse plugin. However I had to
suppress a few due to the simpleJSON lib. I moved the whole lib into a
jar, but its a pain to use due to the old java 1.4 syntax.

However there are more important things to do now and i don't want to
exchange the JSON lib now.